[FEAT/#346] 구독 연장 여부 dialog 로직 연결
