+++ app/src/main/java/com/el/yello/presentation/pay/PayResubsNoticeDialog.kt
@@ -6,58 +6,22 @@ import android.os.Bundle
 import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
-import androidx.core.view.isVisible
-import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.flowWithLifecycle
-import androidx.lifecycle.lifecycleScope
 import com.el.yello.R
 import com.el.yello.databinding.FragmentNoticeResubscribeBinding
-import com.el.yello.util.context.yelloSnackbar
 import com.example.ui.base.BindingDialogFragment
-import com.example.ui.view.UiState
 import com.example.ui.view.setOnSingleClickListener
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
 
 class PayResubsNoticeDialog :
     BindingDialogFragment<FragmentNoticeResubscribeBinding>(R.layout.fragment_notice_resubscribe) {
 
-    private val viewModel by activityViewModels<PayViewModel>()
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        viewModel.getUserSubsInfoStateFromServer()
-        observeSubsNeededState()
-        setNoticeBtnCLickListener()
+        setNoticeBtnClickListener()
     }
 
     override fun onResume() {
         super.onResume()
         showDialogFullScreen()
-    }
-
-    private fun observeSubsNeededState() {
-        viewModel.getUserSubsInfoState.flowWithLifecycle(lifecycle).onEach { state ->
-            when (state) {
-                is UiState.Success -> {
-                    if (state.data?.subscribe.toString() == "canceled") {
-                        view?.isVisible = true
-                        binding.tvResubscribeExpiredDate.text = state.data?.expiredDate
-                    }
-                }
-
-                is UiState.Failure -> {
-                    yelloSnackbar(binding.root, getString(R.string.msg_error))
-                }
-
-                is UiState.Empty -> {
-                    yelloSnackbar(binding.root, getString(R.string.msg_error))
-                }
-
-                is UiState.Loading -> {
-                }
-            }
-        }.launchIn(lifecycleScope)
     }
 
     private fun showDialogFullScreen() {
@@ -69,12 +33,31 @@ class PayResubsNoticeDialog :
         dialog?.window?.clearFlags(WindowManager.LayoutParams.FLAG_DIM_BEHIND)
     }
 
-    private fun setNoticeBtnCLickListener() {
+    private fun setNoticeBtnClickListener() {
         binding.btnNoticeQuit.setOnSingleClickListener {
             dismiss()
         }
         binding.btnYelloplusSubscribe.setOnSingleClickListener {
-            // TODO : 계속 옐로 플러스 구독하기 뷰 이동
+            // TODO: 계속 옐로 플러스 구독하기 뷰 이동
+        }
+    }
+
+    fun setExpiredDate(expiredDate: String) {
+        if (isAdded) {
+            binding.tvResubscribeExpiredDate?.text = expiredDate
+        }
+    }
+
+    companion object {
+        private const val ARG_EXPIRED_DATE = "arg_expired_date"
+
+        @JvmStatic
+        fun newInstance(expiredDate: String): PayResubsNoticeDialog {
+            return PayResubsNoticeDialog().apply {
+                arguments = Bundle().apply {
+                    putString(ARG_EXPIRED_DATE, expiredDate)
+                }
+            }
         }
     }
 }
